2S-2025/Semana8/PC/Consumer/requirements.txt
mysql-connector==2.2.9
mysql-connector-python==9.4.0
paho-mqtt==2.1.0
